优化.
优化.

+++ BGFMDB/libs/BG/BGDB.h
@@ -221,7 +221,7 @@
 /**
  直接调用sqliteb的原生函数计算sun,min,max,avg等.
  */
--(NSInteger)sqliteMethodForTable:(NSString* _Nonnull)name type:(bg_sqliteMethodType)methodType key:(NSString* _Nonnull)key where:(NSString* _Nullable)where;
+-(double)sqliteMethodForTable:(NSString* _Nonnull)name type:(bg_sqliteMethodType)methodType key:(NSString* _Nonnull)key where:(NSString* _Nullable)where;
 /**
  刷新数据库，即将旧数据库的数据复制到新建的数据库,这是为了去掉没用的字段.
  @name 表名称.
